Initial fitting intensities are adjusted to the max of the spectral intensity
As a consequence, amplitudes in input_parameters.xlsx are expressed in
terms of relative rations of each peak.

This requires the amplitudes in older imput_parameters.xlsx to be changed
in order to be compatible with the new version of multifit.

Also, the initial fitting curve can be plotted or not through a flag at the
beginning of the script.